+++ codeforces_core/__init__.py
@@ -2,5 +2,8 @@ __all__ = (
     "interfaces",
     "account",
     "submit",
+    "contest_list",
+    "contest_register",
+    "contest_meta",
 )
 __version__ = "0.0.1"