+++ codeforces_core/constants.py
@@ -1,1 +1,3 @@
 CF_HOST = 'https://codeforces.com'
+# TODO singleton
+CONTESTS_HOME_URL = f"/contests"