+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add auto test(e2e, unit, coverage report)
 - Add vscode-docker container support
 - Reconstruct submit, separate the function as submit / websocket listening / polling submission page
+- Contest register
 
 ### Features